commit of cruft files during development
I've been engaging the repository for a while without committing, so
here's a commit of all the cruft files for the git history.

I've been exploring what kind of diff backend could be used for
asymmetrical streaming output. There don't seem to be good options! But
I'm settling on using xdiff from git and libgit2, and plan to remove
most of the other cruft. Using the same diff code as git brings the
implementation side nearest to what produces expected output.

+++ cli/chat.cpp
@@ -32,7 +32,7 @@ int main([[maybe_unused]]int argc, [[maybe_unused]]char **argv) {
                 break;
             } else while ((extra = cin.rdbuf()->in_avail()) > 1) {
                 size_t msg_size = msg.size();
-                msg.resize(msg_size + extra);
+                msg.resize(msg_size + (size_t)extra);
                 cin.read(&msg[msg_size], extra);
             }
             if (msg.back() == '\n') {
